requirement.txt
flask
pandas
utillc
python-dotenv
opencv-python
scikit-image
scikit-learn